add RemoteAgent
Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

Implements an agent type, `RemoteAgent` that allows for running agents
that run remotely and are accessible via a URL.

To use, declare `.spec.framework: remote` and `.spec.url: <some-url>`
in the agent YAML file.

Note that as of now, the other `.spec` values still need to be
provided but are not used by the RemoteAgent.